need to commit before fetching new branch
